requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
webrtcvad>=2.0.10
openai-whisper>=20240930
pyyaml>=6.0
